// src/app/common/constans/models/IUserFilters.ts
export interface IUserFilters {
  ordercolumn: string | null;
  orderdirection: string| null;
  pagenumber: number| null;
  pagesize: number| null;
  identificacion: string| null;
  id_usuario: number| null;
  tipo_persona: number| null;
  tipo_identificacion: number| null;
  razon_social: string| null;
  nombre: string| null;
  email: string| null;
  genero: number| null;
  ciudad: number| null;
  telefono: string| null;
  id_rol: number| null;
  estado: number| null;
  fechainicio: string| null;
  fechafinal: string| null;
}